refactor pipeline file and tests to match
